refactor: 💡 Updates based on PR  comments
