+++ package.json
@@ -47,6 +47,7 @@
     "@types/dompurify": "^2.4.0",
     "@types/jest": "^29.5.12",
     "@types/lodash": "^4.14.191",
+    "@types/matter-js": "^0.19.6",
     "@types/node": "18.11.18",
     "@types/react": "18.0.27",
     "@types/react-dom": "18.0.10",
